+++ packages/databricks-vscode/src/configuration/auth/MetadataService.test.ts
@@ -10,7 +10,6 @@ import {
     MetadataServiceVersionHeader,
     RequestVisitor,
     logging,
-    fetch,
 } from "@databricks/databricks-sdk";
 const {NamedLogger} = logging;
 
@@ -50,7 +49,7 @@ describe(__filename, function () {
                         name: "pat",
                         async configure(): Promise<RequestVisitor> {
                             return async (headers: Headers) => {
-                                headers["Authorization"] = `Bearer XXXX`;
+                                headers.set("Authorization", "Bearer XXXX");
                             };
                         },
                     },
@@ -93,7 +92,7 @@ describe(__filename, function () {
                         name: "pat",
                         async configure(): Promise<RequestVisitor> {
                             return async (headers: Headers) => {
-                                headers["Authorization"] = `Bearer XXXX`;
+                                headers.set("Authorization", "Bearer XXXX");
                             };
                         },
                     },
@@ -109,9 +108,9 @@ describe(__filename, function () {
 
         const apiClient = new ApiClient(config);
 
-        const headers: Record<string, string> = {};
+        const headers = new Headers();
         await apiClient.config.authenticate(headers);
 
-        assert.equal(headers["Authorization"], "Bearer XXXX");
+        assert.equal(headers.get("Authorization"), "Bearer XXXX");
     });
 });