+++ packages/databricks-vscode/src/extension.ts
@@ -34,7 +34,7 @@ import {
 } from "./workspace-fs";
 import {registerBundleAutocompleteProvider} from "./bundle/bundleAutocompleteProvider";
 import {CustomWhenContext} from "./vscode-objs/CustomWhenContext";
-import {WorkspaceStateManager} from "./vscode-objs/WorkspaceState";
+import {StateStorage} from "./vscode-objs/StateStorage";
 import path from "node:path";
 import {MetadataServiceManager} from "./configuration/auth/MetadataServiceManager";
 import {FeatureId, FeatureManager} from "./feature-manager/FeatureManager";
@@ -52,6 +52,7 @@ import {NotebookInitScriptManager} from "./language/notebooks/NotebookInitScript
 import {showRestartNotebookDialogue} from "./language/notebooks/restartNotebookDialogue";
 import {BundleWatcher} from "./file-managers/BundleWatcher";
 import {BundleFileSet} from "./bundle/BundleFileSet";
+import {showWhatsNewPopup} from "./whatsNewPopup";
 
 export async function activate(
     context: ExtensionContext
@@ -86,7 +87,7 @@ export async function activate(
         return undefined;
     }
 
-    const workspaceStateManager = new WorkspaceStateManager(context);
+    const stateStorage = new StateStorage(context);
 
     // Add the databricks binary to the PATH environment variable in terminals
     context.environmentVariableCollection.clear();
@@ -121,7 +122,7 @@ export async function activate(
     const pythonExtensionWrapper = new MsPythonExtensionWrapper(
         pythonExtension,
         workspace.workspaceFolders[0].uri,
-        workspaceStateManager
+        stateStorage
     );
 
     context.subscriptions.push(
@@ -145,7 +146,7 @@ export async function activate(
 
     // Configuration group
     const cli = new CliWrapper(context);
-    const connectionManager = new ConnectionManager(cli, workspaceStateManager);
+    const connectionManager = new ConnectionManager(cli, stateStorage);
     context.subscriptions.push(
         connectionManager.onDidChangeState(async (state) => {
             telemetry.setMetadata(
@@ -172,7 +173,7 @@ export async function activate(
     );
     const workspaceFsCommands = new WorkspaceFsCommands(
         workspace.workspaceFolders[0].uri,
-        workspaceStateManager,
+        stateStorage,
         connectionManager,
         workspaceFsDataProvider
     );
@@ -210,7 +211,7 @@ export async function activate(
 
     const wsfsAccessVerifier = new WorkspaceFsAccessVerifier(
         connectionManager,
-        workspaceStateManager,
+        stateStorage,
         synchronizer,
         telemetry
     );
@@ -218,12 +219,10 @@ export async function activate(
     context.subscriptions.push(wsfsAccessVerifier);
 
     const dbConnectInstallPrompt = new DbConnectInstallPrompt(
-        workspaceStateManager,
+        stateStorage,
         pythonExtensionWrapper
     );
-    const featureManager = new FeatureManager<FeatureId>([
-        "notebooks.dbconnect",
-    ]);
+    const featureManager = new FeatureManager<FeatureId>([]);
     featureManager.registerFeature(
         "debugging.dbconnect",
         () =>
@@ -240,7 +239,7 @@ export async function activate(
             new NotebookAccessVerifier(
                 featureManager,
                 pythonExtensionWrapper,
-                workspaceStateManager
+                stateStorage
             )
     );
 
@@ -323,7 +322,7 @@ export async function activate(
 
     const configureAutocomplete = new ConfigureAutocomplete(
         context,
-        workspaceStateManager,
+        stateStorage,
         workspace.workspaceFolders[0].uri.fsPath,
         pythonExtensionWrapper,
         dbConnectInstallPrompt
@@ -340,7 +339,7 @@ export async function activate(
     const configurationDataProvider = new ConfigurationDataProvider(
         connectionManager,
         synchronizer,
-        workspaceStateManager,
+        stateStorage,
         wsfsAccessVerifier,
         featureManager,
         telemetry
@@ -572,6 +571,18 @@ export async function activate(
             e
         );
     });
+
+    showWhatsNewPopup(context, stateStorage)
+        .catch((e) => {
+            logging.NamedLogger.getOrCreate(Loggers.Extension).error(
+                "Error while showing popup for what's new",
+                e
+            );
+        })
+        .finally(() => {
+            stateStorage.lastInstalledExtensionVersion =
+                packageMetadata.version;
+        });
 
     CustomWhenContext.setActivated(true);
     telemetry.recordEvent(Events.EXTENSION_ACTIVATED);