+++ packages/databricks-vscode/src/test/e2e/run_job_on_cluster_with_repo.e2e.ts
@@ -126,9 +126,9 @@ describe("Run job on cluster with repo", async function () {
                     await browser.getTextByLabel(labelToDefaults.status.label)
                 ).match(/Succeeded/) !== null,
             {
-                timeout: 20000,
-                interval: 50,
-                timeoutMsg: "Job did not reach succeeded status after 20s.",
+                timeout: 30000,
+                interval: 100,
+                timeoutMsg: "Job did not reach succeeded status after 30s.",
             }
         );
 
@@ -201,9 +201,9 @@ describe("Run job on cluster with repo", async function () {
                     await browser.getTextByLabel(labelToDefaults.status.label)
                 ).includes("Succeeded"),
             {
-                timeout: 20000,
+                timeout: 30000,
                 interval: 100,
-                timeoutMsg: "Job did not reach succeeded status after 20s.",
+                timeoutMsg: "Job did not reach succeeded status after 30s.",
             }
         );